Use `setq-default` for showing trailing whitespace
Some variables in Emacs are "buffer-local", meaning that each buffer is
allowed to have a separate value for that variable that overrides the
global default.

If a variable is buffer-local, then `setq` sets its local value in the
current buffer and `setq-default` sets the global default value.

If a variable is not buffer-local, then `setq` and `setq-default` do the
same thing.

In this case, we want to use `setq-default` as we always want to show
trailing whitespace.